Try to tune SmallVectorSizeType 3

+++ llvm/include/llvm/ADT/SmallVector.h
@@ -107,7 +107,7 @@ protected:
 template <class T>
 using SmallVectorSizeType =
     std::conditional_t<sizeof(T) < 4 && sizeof(void *) >= 8, uint64_t,
-                       uint32_t>;
+                       uint64_t>;
 
 /// Figure out the offset of the first element.
 template <class T, typename = void> struct SmallVectorAlignmentAndSize {
